Added editor UI to configure IDs
Added an option in the editor under UWP / Xbox Live / Configure IDs to pop up a window that will load and save your SCID and Title IDs for you as an asset file. This asset file is then used when building to auto generate and add the config file to your VS project.